refactor(steps): move shared binding helpers to apply_pvyml_services
The four helpers (findServiceByName, bindProjectService,
bindProjectPostgres, bindProjectMysql) live in detect_services.go
today but are also used by ApplyPvYmlServicesStep. Moving them now
ahead of PR 5's deletion of the auto-detect step so the surviving
caller keeps working. No behavior change.
