+++ internal/automation/steps/apply_pvyml_services.go
@@ -7,6 +7,7 @@ import (
 	"github.com/prvious/pv/internal/mysql"
 	"github.com/prvious/pv/internal/postgres"
 	"github.com/prvious/pv/internal/redis"
+	"github.com/prvious/pv/internal/registry"
 )
 
 // ApplyPvYmlServicesStep binds the services declared in a project's
@@ -84,4 +85,60 @@ func (s *ApplyPvYmlServicesStep) Run(ctx *automation.Context) (string, error) {
 	}
 
 	return fmt.Sprintf("bound %d service(s) from pv.yml", count), nil
+}
+
+func findServiceByName(reg *registry.Registry, name string) string {
+	for key := range reg.Services {
+		keyName, _ := registry.ParseServiceKey(key)
+		if keyName == name {
+			return key
+		}
+	}
+	return ""
+}
+
+func bindProjectService(reg *registry.Registry, projectName, svcType, svcKey string) {
+	for i := range reg.Projects {
+		if reg.Projects[i].Name != projectName {
+			continue
+		}
+		if reg.Projects[i].Services == nil {
+			reg.Projects[i].Services = &registry.ProjectServices{}
+		}
+		switch svcType {
+		case "redis":
+			reg.Projects[i].Services.Redis = true
+		case "mail":
+			reg.Projects[i].Services.Mail = true
+		case "s3":
+			reg.Projects[i].Services.S3 = true
+		}
+		break
+	}
+}
+
+func bindProjectPostgres(reg *registry.Registry, projectName, major string) {
+	for i := range reg.Projects {
+		if reg.Projects[i].Name != projectName {
+			continue
+		}
+		if reg.Projects[i].Services == nil {
+			reg.Projects[i].Services = &registry.ProjectServices{}
+		}
+		reg.Projects[i].Services.Postgres = major
+		return
+	}
+}
+
+func bindProjectMysql(reg *registry.Registry, projectName, version string) {
+	for i := range reg.Projects {
+		if reg.Projects[i].Name != projectName {
+			continue
+		}
+		if reg.Projects[i].Services == nil {
+			reg.Projects[i].Services = &registry.ProjectServices{}
+		}
+		reg.Projects[i].Services.MySQL = version
+		return
+	}
 }
